backend/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.6.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
cryptography==3.4.8
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.7
django-allauth==0.45.0
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
joblib==1.0.1
JPype1==1.0.2
konlpy==0.5.2
lxml==4.6.3
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.3
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sqlparse==0.4.1
threadpoolctl==2.2.0
toml==0.10.2
tweepy==3.7.0
urllib3==1.26.6
